dreamdiffusion/requirements.txt
einops==0.3.0
kornia==0.7.0
matplotlib==3.7.4
natsort==8.4.0
numpy==1.24.4
omegaconf==2.1.1
packaging==23.2
Pillow==9.0.1
pandas
pytorch_lightning==1.6.5
scipy
setuptools==68.0.0
scikit-learn
timm==0.5.4
torch==1.12.1+cu116
torchmetrics==0.9.2
torchvision==0.13.1+cu116
tqdm==4.64.0
transformers==4.35.2
wandb==0.12.21
google-api-core==2.14.0
google-auth==2.24.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
